fix: docker layer caching with docker buildx (#249)
* feat: added validation to use docker context

* fix: add --context flag validation

* fix: addressed shellcheck errors

* fix: added warning for multi-arch builds

* chore: added new lines to dlcwarning

* chore: fixed typo

* chore: addressed pr comments

* chore: refactored code for less repetition

* refactor: expand parameters

* chore: code cleanup

* feat: set-repo-policy command (#250)

* chore: renamed file

* feat: add set-repository-permissions (#239)

* fix: corrected file name

* ci: add test for set-repo-policy command

* fix: add valid repo-policy

+++ src/scripts/docker-buildx.sh
@@ -8,7 +8,9 @@ ORB_EVAL_PUBLIC_REGISTRY_ALIAS=$(eval echo "${ORB_EVAL_PUBLIC_REGISTRY_ALIAS}")
 ORB_EVAL_EXTRA_BUILD_ARGS=$(eval echo "${ORB_EVAL_EXTRA_BUILD_ARGS}")
 ECR_COMMAND="ecr"
 number_of_tags_in_ecr=0
-docker_tag_args=""
+
+IFS=', ' read -ra platform <<<"${ORB_VAL_PLATFORM}"
+number_of_platforms="${#platform[@]}"
 
 if [ -z "${!ORB_ENV_REGISTRY_ID}" ]; then
   echo "The registry ID is not found. Please add the registry ID as an environment variable in CicleCI before continuing."
@@ -47,24 +49,36 @@ if [ "${ORB_VAL_SKIP_WHEN_TAGS_EXIST}" = "0" ] || [[ "${ORB_VAL_SKIP_WHEN_TAGS_E
   fi
 
   if [ -n "${ORB_EVAL_EXTRA_BUILD_ARGS}" ]; then
-    ORB_EVAL_EXTRA_BUILD_ARGS=$(eval echo "${ORB_EVAL_EXTRA_BUILD_ARGS}")
     set -- "$@" "${ORB_EVAL_EXTRA_BUILD_ARGS}"
   fi
 
-  if ! docker context ls | grep builder; then
-    # We need to skip the creation of the builder context if it's already present
-    # otherwise the command will fail when called more than once in the same job.
 
-    docker context create builder
-    docker run --privileged --rm tonistiigi/binfmt --install all
-    docker --context builder buildx create --use
-  fi
 
-  docker --context builder buildx build \
+  if [ "${number_of_platforms}" -gt 1 ]; then
+    # In order to build multi-architecture images, a context with binfmt installed must be used. 
+    # However, Docker Layer Caching with multi-architecture builds is not currently supported
+
+    if ! docker context ls | grep builder; then
+      # We need to skip the creation of the builder context if it's already present
+      # otherwise the command will fail when called more than once in the same job.
+      docker context create builder
+      docker run --privileged --rm tonistiigi/binfmt --install all
+      docker --context builder buildx create --use
+    fi
+    context_args="--context builder"
+    echo -e "\n \n WARNING: Docker Layer Caching is currently not supported for multi-architecture image builds. \n \n"
+  fi 
+  
+  set -x
+  docker \
+    ${context_args:+$context_args} \
+    buildx build \
     -f "${ORB_EVAL_PATH}"/"${ORB_VAL_DOCKERFILE}" \
-    ${docker_tag_args} \
+    ${docker_tag_args:+$docker_tag_args} \
     --platform "${ORB_VAL_PLATFORM}" \
     --progress plain \
     "$@" \
     "${ORB_EVAL_PATH}"
+  set +x
+  
 fi
